# docs/swagger.yaml
basePath: /v1
definitions:
...
  routes.HoldParams:
    properties:
      amount:
        description: original hold amount
        type: string
      id:
        type: string
      reason_code:
        type: string
      transfer_id:
        type: string
    type: object
  routes.HoldResponse:
    properties:
      account_id:
        type: string
      amount:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: original hold amount
      captured_amount:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: converted to real debit
      captured_at:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      reason_code:
        type: string
      released_amount:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: returned to available
      released_at:
        type: string
      remaining_amount:
        allOf:
        - $ref: '#/definitions/routes.MoneyAmount'
        description: still locked
      status:
        description: '"active" | "consumed" | "released" | "expired"'
        type: string
    type: object
  routes.MoneyAmount:
    properties:
      amount:
        description: always a string — never float for money
        type: string
      currency:
        description: '"NGN"'
        type: string
    type: object
...
  routes.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      external_ref:
        type: string
      full_name:
        type: string
      id:
        type: string
      national_id:
        type: string
      phone:
        type: string
      status:
        type: string
    type: object
...
  routes.errResponse:
    properties:
      error:
        type: string
    type: object
  routes.parameters:
    properties:
      email:
        type: string
      external_ref:
        type: string
      full_name:
        type: string
      id:
        type: string
      idemp_key:
        type: string
      limit:
        type: integer
      metadata:
        additionalProperties: true
        type: object
      national_id:
        type: string
      offset:
        type: integer
      phone:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A comprehensive distributed payment processing system with support
    for transfers, deposits, withdrawals, and holds.
  title: Distributed Payments Engine API
  version: 1.0.0
paths:
...
  /admin/hold/{transfer_id}:
    get:
      description: Retrieves the active hold placed on a transfer for payment processing.
      parameters:
      - description: Transfer request UUID
        format: uuid
        in: path
        name: transfer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.HoldResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Get active hold for transfer
      tags:
      - Admin
  /admin/hold/consume:
    post:
      consumes:
      - application/json
      description: Converts a held amount to an actual debit and releases remaining
        funds.
      parameters:
      - description: Hold Consumption Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.HoldParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Consume a hold
      tags:
      - Admin
...
  /create/user:
    post:
      consumes:
      - application/json
      description: Creates a new customer account with the provided information. Idempotent
        by external_ref.
      parameters:
      - description: User Creation Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.parameters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Create a new customer/user
      tags:
      - Users
...
  /list/users:
    post:
      consumes:
      - application/json
      description: Returns paginated list of customers with optional limit and offset.
      parameters:
      - description: Pagination details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.parameters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: List customers with pagination
      tags:
      - Users
...
  /update/user:
    post:
      consumes:
      - application/json
      description: Updates the status of an existing customer (e.g., active, inactive,
        suspended).
      parameters:
      - description: User Status Update Details (requires id and status)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.parameters'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Update user status
      tags:
      - Users
  /user/{id}:
    get:
      description: Retrieves a specific customer by their UUID.
      parameters:
      - description: Customer UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Get user by ID
      tags:
      - Users
...
swagger: "2.0"