# src/routes/handler_user.go
package routes

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	database "github.com/OMODON-ETEMI/distributed-payments-engine/src/database/gen"
	"github.com/go-chi/chi"
	"github.com/jackc/pgx/v5"
	"github.com/jackc/pgx/v5/pgtype"
)

type parameters struct {
	IdempKey    string                 `json:"idemp_key"`
	ID          string                 `json:"id"`
	ExternalRef string                 `json:"external_ref"`
	FullName    string                 `json:"full_name"`
	Email       string                 `json:"email"`
	Phone       string                 `json:"phone"`
	NationalID  string                 `json:"national_id"`
	Status      string                 `json:"status"`
	Metadata    map[string]interface{} `json:"metadata"`
	Limit       int                    `json:"limit"`
	Offset      int                    `json:"offset"`
}

// HandleCreateUser creates a new customer/user.
// @Summary Create a new customer/user
// @Description Creates a new customer account with the provided information. Idempotent by external_ref.
// @Tags Users
// @Accept json
// @Produce json
// @Param body body parameters true "User Creation Details"
// @Success 200 {object} UserResponse
// @Failure 400 {object} errResponse
// @Failure 500 {object} errResponse
// @Router /create/user [post]
func (api *ApiConfig) HandleCreateUser(w http.ResponseWriter, r *http.Request) {
	decoder := json.NewDecoder(r.Body)
	params := parameters{}
	err := decoder.Decode(&params)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing Json: %v", err))
		return
	}

	if params.ExternalRef == "" || params.FullName == "" || params.Email == "" {
		respondWithError(w, 400, "missing required fields: external_ref, full_name or email")
		return
	}

	existing, err := api.Db.Queries.GetCustomerByExternalRef(r.Context(), params.ExternalRef)
	if err != nil && !errors.Is(err, pgx.ErrNoRows) {
		respondWithError(w, 500, fmt.Sprintf("Internal Server Error: %v", err))
		return
	}
	if err == nil {
		respondeWithJson(w, 200, UserResponseObject(existing))
		return
	}
	if params.Metadata == nil {
		params.Metadata = make(map[string]interface{})
	}
	metadataBytes, err := json.Marshal(params.Metadata)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing metadata: %v", err))
		return
	}

	user, err := api.Db.Queries.CreateCustomer(r.Context(), database.CreateCustomerParams{
		ExternalRef: params.ExternalRef,
		FullName:    params.FullName,
		Email:       params.Email,
		Phone:       pgtype.Text{String: params.Phone, Valid: params.Phone != ""},
		NationalID:  pgtype.Text{String: params.NationalID, Valid: params.NationalID != ""},
		Status:      params.Status,
		Metadata:    metadataBytes,
	})
	if err != nil {
		respondWithError(w, 500, fmt.Sprintf("Error creating Customer: %v", err))
		return
	}
	respondeWithJson(w, 200, UserResponseObject(user))

}

func (api *ApiConfig) HandleGetUserByExternalRef(w http.ResponseWriter, r *http.Request) {
	decoder := json.NewDecoder(r.Body)
	params := parameters{}
	err := decoder.Decode(&params)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing metadata: %v", err))
		return
	}
	if params.ExternalRef == "" {
		respondWithError(w, 400, "external_ref is required")
		return
	}
	user, err := api.Db.Queries.GetCustomerByExternalRef(r.Context(), params.ExternalRef)
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			respondWithError(w, 404, "user not found")
			return
		}
		respondWithError(w, 500, fmt.Sprintf("Error getting user by External Ref: %v", err))
		return
	}
	respondeWithJson(w, 200, UserResponseObject(user))
}

// HandleGetUserById retrieves a specific customer by their UUID.
// @Summary Get user by ID
// @Description Retrieves a specific customer by their UUID.
// @Tags Users
// @Produce json
// @Param id path string true "Customer UUID" format(uuid)
// @Success 200 {object} UserResponse
// @Failure 400 {object} errResponse
// @Failure 404 {object} errResponse
// @Failure 500 {object} errResponse
// @Router /user/{id} [get]
func (api *ApiConfig) HandleGetUserById(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	if idStr == "" {
		respondWithError(w, 400, "User Id is required")
		return
	}
	id, err := StringtoPgUuid(idStr)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing ID: %v", err))
		return
	}
	user, err := api.Db.Queries.GetCustomerByID(r.Context(), id)
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			respondWithError(w, 404, "user not found")
			return
		}
		respondWithError(w, 500, fmt.Sprintf("Error getting user by ID: %v", err))
		return
	}
	respondeWithJson(w, 200, UserResponseObject(user))
}

// HandleUserUpdateStatus updates the status of an existing customer.
// @Summary Update user status
// @Description Updates the status of an existing customer (e.g., active, inactive, suspended).
// @Tags Users
// @Accept json
// @Produce json
// @Param body body parameters true "User Status Update Details (requires id and status)"
// @Success 200 {object} UserResponse
// @Failure 400 {object} errResponse
// @Failure 404 {object} errResponse
// @Failure 500 {object} errResponse
// @Router /update/user [post]
func (api *ApiConfig) HandleUserUpdateStatus(w http.ResponseWriter, r *http.Request) {
	decoder := json.NewDecoder(r.Body)
	params := parameters{}
	err := decoder.Decode(&params)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing metadata: %v", err))
		return
	}
	id, err := StringtoPgUuid(params.ID)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing ID: %v", err))
		return
	}
	if params.Status == "" {
		respondWithError(w, 400, "status is required")
		return
	}
	user, err := api.Db.Queries.UpdateCustomerStatus(r.Context(), database.UpdateCustomerStatusParams{
		ID:     id,
		Status: params.Status,
	})
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			respondWithError(w, 404, "user not found")
			return
		}
		respondWithError(w, 500, fmt.Sprintf("Error updating user status: %v", err))
		return
	}
	respondeWithJson(w, 200, UserResponseObject(user))
}

// HandleListCustomers returns a paginated list of customers.
// @Summary List customers with pagination
// @Description Returns paginated list of customers with optional limit and offset.
// @Tags Users
// @Accept json
// @Produce json
// @Param body body parameters true "Pagination details"
// @Success 200 {array} UserResponse
// @Failure 400 {object} errResponse
// @Failure 500 {object} errResponse
// @Router /list/users [post]
func (api *ApiConfig) HandleListCustomers(w http.ResponseWriter, r *http.Request) {
	decoder := json.NewDecoder(r.Body)
	params := parameters{}
	err := decoder.Decode(&params)
	if err != nil {
		respondWithError(w, 400, fmt.Sprintf("Error parsing Json: %v", err))
		return
	}

	limit := int32(50)
	offset := int32(0)
	if params.Limit > 0 {
		limit = int32(params.Limit)
	}
	if params.Offset > 0 {
		offset = int32(params.Offset)
	}
	if limit > 1000 {
		limit = 1000
	}

	users, err := api.Db.Queries.ListCustomers(r.Context(), database.ListCustomersParams{Limit: limit, Offset: offset})
	if err != nil {
		respondWithError(w, 500, fmt.Sprintf("Error listing users: %v", err))
		return
	}

	var out []UserResponse
	for _, u := range users {
		out = append(out, UserResponseObject(u))
	}
	respondeWithJson(w, 200, out)
}

// {
//   "id": "61fd6667-9a2e-476e-a4f2-733f205d8c52",
//   "external_ref": "ext_ref_77889",
//   "full_name": "Oritsetemi Glory Omodon",
//   "email": "glory@example.com",
//   "phone": "+2348012345678",
//   "national_id": "NIN-123456789",
//   "status": "active",
//   "metadata": "eyJyaXNrX3Njb3JlIjogMC41LCAicmVmZXJyZWRfYnkiOiAic3RhcnR1cF9hbHBoYSIsICJvbmJvYXJkaW5nX3NvdXJjZSI6ICJ3ZWJfYXBwIn0=",
//   "created_at": "2026-04-16T18:12:20.821268Z",
//   "updated_at": "2026-04-16T18:12:20.821268Z",
//   "deleted_at": null
// }

// {
//   "idemp_key": "user_create_9921_abc",
//   "external_ref": "ext_ref_77889",
//   "full_name": "Oritsetemi Glory Omodon",
//   "email": "glory@example.com",
//   "phone": "+2348012345678",
//   "national_id": "NIN-123456789",
//   "status": "active",
//   "metadata": {
//     "onboarding_source": "web_app",
//     "referred_by": "startup_alpha",
//     "risk_score": 0.5
//   }
// }